testing separated into relevant files; fixed null reference warnings
