+++ VacationApp/VacationApp.Tests/cobertura/index.html
@@ -33,7 +33,7 @@
 </tr>
 <tr>
 <th>Coverage date:</th>
-<td class="limit-width " title="04/23/2025 - 22:13:42">04/23/2025 - 22:13:42</td>
+<td class="limit-width " title="05/01/2025 - 04:47:30">05/01/2025 - 04:47:30</td>
 </tr>
 </table>
 </div>
@@ -291,7 +291,7 @@
 </table>
 </div>
 </coverage-info>
-<div class="footer">Generated by: ReportGenerator 5.4.5.0<br />04/23/2025 - 22:19:23<br /><a href="https://github.com/danielpalme/ReportGenerator">GitHub</a> | <a href="https://reportgenerator.io">reportgenerator.io</a></div></div></div>
+<div class="footer">Generated by: ReportGenerator 5.4.5.0<br />05/01/2025 - 04:47:49<br /><a href="https://github.com/danielpalme/ReportGenerator">GitHub</a> | <a href="https://reportgenerator.io">reportgenerator.io</a></div></div></div>
 <script type="text/javascript">
 /* <![CDATA[ */
 (function() {
